refactor: centralize shared configs and security policies
