+++ setup-service/src/main/java/com/ejada/setup/security/SetupAuthorized.java
@@ -1,7 +1,6 @@
 package com.ejada.setup.security;
 
-import org.springframework.security.access.prepost.PreAuthorize;
-
+import com.ejada.starter_security.authorization.PlatformStaffAuthorized;
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -11,9 +10,6 @@ import java.lang.annotation.Target;
 @Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@PreAuthorize(
-        "@roleChecker.hasRole(authentication, T(com.ejada.starter_security.Role).EJADA_OFFICER, "
-                + "T(com.ejada.starter_security.Role).TENANT_ADMIN, "
-                + "T(com.ejada.starter_security.Role).TENANT_OFFICER)")
+@PlatformStaffAuthorized
 public @interface SetupAuthorized {
 }
