+++ shared-lib/shared-config/src/main/resources/application-shared-tenant-qa.yaml
@@ -1,19 +1,21 @@
+app:
+  environment: qa
+  service-code: tenant
+  schema: tenant
+
 spring:
-  config:
-    import: optional:classpath:application-shared-platform-dev.yaml
   datasource:
-    url: "${DB_URL:jdbc:postgresql://${DB_HOST:localhost}:${DB_PORT:5432}/${DB_NAME:lms}?currentSchema=catalog}"
-    username: ${DB_USERNAME:postgres}
-    password: ${DB_PASSWORD:postgres}
+    url: "${DB_URL:jdbc:postgresql://${DB_HOST:localhost}:${DB_PORT:5432}/${DB_NAME:ejada}?currentSchema=${app.schema}}"
+    username: ${DB_USERNAME}
+    password: ${DB_PASSWORD}
     driver-class-name: org.postgresql.Driver
     hikari:
-      pool-name: HikariPool-ejada
+      pool-name: HikariPool-Ejada
       maximum-pool-size: 10
       minimum-idle: 2
       idle-timeout: 30000        # 30s
       connection-timeout: 300000 # 300s
       max-lifetime: 1800000      # 30m
-
   jpa:
     open-in-view: false
     show-sql: true
@@ -21,49 +23,38 @@ spring:
       ddl-auto: none
     properties:
       hibernate:
-        default_schema: catalog
+        ddl-auto: none
+        default_schema: ${app.schema}
         hbm2ddl.create_namespaces: true
         format_sql: true
         jdbc.time_zone: UTC
-
   flyway:
     enabled: true
     baseline-on-migrate: true
     locations: classpath:db/migration/common,classpath:db/migration/{vendor}
-    schemas: public,catalog
-    default-schema: catalog
-
+    schemas: public,${app.schema}
+    default-schema: ${app.schema}
   kafka:
-    client-id: ejada-catalog-dev
-
+    bootstrap-servers: ${KAFKA_BOOTSTRAP_SERVERS:kafka:9092}
+    client-id: ${app.kafka-client-id:ejada-${app.service-code}-${app.environment}}
+    consumer:
+      group-id: ${KAFKA_CONSUMER_GROUP:ejada-backend-${app.environment}}
+      auto-offset-reset: earliest
+      enable-auto-commit: false
+      properties:
+        max.poll.interval.ms: 300000
+        max.poll.records: 500
+    producer:
+      acks: all
+      retries: 3
+      batch-size: 16384
+      compression-type: gzip
+      properties:
+        linger.ms: 5
   data:
     redis:
       repositories:
         enabled: false
-
-management:
-  endpoints:
-    web:
-      base-path: /actuator
-      exposure:
-        include: health,info,metrics,prometheus,beans,conditions,configprops,loggers,threaddump
-  endpoint:
-    health:
-      show-details: always
-  tracing:
-    sampling:
-      probability: 1.0
-
-springdoc:
-  api-docs:
-    enabled: true
-    path: /v3/api-docs
-  swagger-ui:
-    enabled: true
-    display-request-duration: true
-    filter: true
-    operationsSorter: method
-    tagsSorter: alpha
 
 shared:
   core:
@@ -111,7 +102,7 @@ shared:
       frame-options: SAMEORIGIN
       content-type-options: nosniff
       referrer-policy: no-referrer
-      xss-protection: "0"
+      xss-protection: "1; mode=block"
     propagation:
       enabled: true
       include:
@@ -142,7 +133,7 @@ shared:
     sinks:
       db:
         enabled: true
-        schema: catalog
+        schema: ${app.schema}
         table: audit_logs
       kafka:
         enabled: true
@@ -150,49 +141,33 @@ shared:
       otlp:
         enabled: false
       outbox:
-        enabled: true
-  openapi:
-    enabled: true
-    title: " catalog Service API"
-    description: "catalog service endpoints for tenant/platform configuration."
-    version: "1.0.0"
-    servers:
-      - http://localhost:8080/core
-    group:
-      name: catalog
-      packages-to-scan:
-        - com.ejada.catalog
-    jwt-security: false
+        enabled: false
   redis:
-    host: redis
-    port: 6379
-    timeout: 2s
-    key-prefix: catalog
-    default-ttl: 600s
+    host: ${REDIS_HOST:redis}
+    port: ${REDIS_PORT:6379}
+    timeout: ${REDIS_TIMEOUT:2s}
+    key-prefix: ejada
+    default-ttl: ${REDIS_DEFAULT_TTL:600s}
     reactive: false
   crypto:
     algorithm: AES_GCM
     in-memory:
-      activeKid: ${CRYPTO_ACTIVE_KID:local-dev-key}
+      activeKid: ${CRYPTO_ACTIVE_KID:local-qa-key}
       keys:
-        local-dev-key: ${CRYPTO_LOCAL_DEV_KEY:4J8bfqUaEqzJCQakoJPM3w==}
-
-logging:
-  level:
-    root: ERROR
-    com.ejada: ERROR
-    com.ejada.audit.starter: ERROR
-    com.ejada.audit.starter.core.dispatch: ERROR
-    com.ejada.audit.starter.core.dispatch.sinks: ERROR
-    org.hibernate.tool.schema: ERROR
-    org.hibernate.SQL: ERROR
-    org.hibernate.orm.jdbc.bind: ERROR
-    org.springframework.cache: ERROR
-    org.springframework.cache.interceptor.CacheAspectSupport: ERROR
-
-app:
-  tenant-provisioning:
-    topic: tenant.provisioning
-    consumer-group: catalog-tenant-provisioning
-
-debug: true
+        local-qa-key: ${CRYPTO_LOCAL_QA_KEY}
+  security:
+    mode: hs256
+    jwt:
+      secret: ${JWT_SECRET}
+      token-period: 15m
+    hs256:
+      secret: ${JWT_SECRET}
+    resource-server:
+      enabled: true
+      permit-all:
+        - "**"
+      disable-csrf: false
+    stateless: true
+    roles-claim: roles
+    tenant-claim: tenant
+    enable-role-check: true
