+++ shared-lib/shared-starters/starter-security/src/main/java/com/ejada/starter_security/authorization/PlatformStaffAuthorized.java
@@ -1,0 +1,21 @@
+package com.ejada.starter_security.authorization;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import org.springframework.security.access.prepost.PreAuthorize;
+
+/**
+ * Grants access to EJADA platform staff (EJADA officer, tenant admin, tenant officer).
+ */
+@Target({ElementType.METHOD, ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@PreAuthorize("@roleChecker.hasRole(authentication, "
+        + "T(com.ejada.starter_security.Role).EJADA_OFFICER, "
+        + "T(com.ejada.starter_security.Role).TENANT_ADMIN, "
+        + "T(com.ejada.starter_security.Role).TENANT_OFFICER)")
+public @interface PlatformStaffAuthorized {
+}
