+++ shared-lib/shared-starters/starter-security/src/main/java/com/ejada/starter_security/authorization/TenantAccessAuthorized.java
@@ -1,15 +1,18 @@
-package com.ejada.sec.security;
+package com.ejada.starter_security.authorization;
 
-import com.ejada.starter_security.authorization.PlatformStaffAuthorized;
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+import org.springframework.security.access.prepost.PreAuthorize;
 
+/**
+ * Delegates authorization to the tenant access policy.
+ */
 @Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@PlatformStaffAuthorized
-public @interface SecAuthorized {
+@PreAuthorize("@tenantAccessPolicy.isAllowed(authentication)")
+public @interface TenantAccessAuthorized {
 }
