+++ tenant-platform/catalog-service/src/main/java/com/ejada/catalog/security/CatalogAuthorized.java
@@ -1,7 +1,6 @@
 package com.ejada.catalog.security;
 
-import org.springframework.security.access.prepost.PreAuthorize;
-
+import com.ejada.starter_security.authorization.EjadaOfficerAuthorized;
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -11,6 +10,6 @@ import java.lang.annotation.Target;
 @Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@PreAuthorize("@roleChecker.hasRole(authentication, T(com.ejada.starter_security.Role).EJADA_OFFICER")
+@EjadaOfficerAuthorized
 public @interface CatalogAuthorized {
 }
