+++ tenant-platform/subscription-service/src/main/resources/application-dev.yaml
@@ -1,4 +1,6 @@
 spring:
+  config:
+    import: optional:classpath:application-shared-platform-dev.yaml
   datasource:
     url: "${DB_URL:jdbc:postgresql://${DB_HOST:localhost}:${DB_PORT:5432}/${DB_NAME:lms}?currentSchema=subscription}"
     username: ${DB_USERNAME:postgres}
@@ -32,22 +34,7 @@ spring:
     default-schema: subscription
 
   kafka:
-    bootstrap-servers: kafka:9092
     client-id: ejada-subscription-dev
-    consumer:
-      group-id: ejada-backend-dev
-      auto-offset-reset: earliest
-      enable-auto-commit: false
-      properties:
-        max.poll.interval.ms: 300000
-        max.poll.records: 500
-    producer:
-      acks: all
-      retries: 3
-      batch-size: 16384
-      compression-type: gzip
-      properties:
-        linger.ms: 5
 
   data:
     redis:
