+++ tenant-platform/tenant-service/src/main/resources/application-qa.yaml
@@ -1,58 +1,14 @@
 spring:
-  datasource:
-    url: "${DB_URL:jdbc:postgresql://${DB_HOST:localhost}:${DB_PORT:5432}/${DB_NAME:ejada}?currentSchema=tenant}"
-    username: ${DB_USERNAME}
-    password: ${DB_PASSWORD}
-    driver-class-name: org.postgresql.Driver
-    hikari:
-      pool-name: HikariPool-Ejada
-      maximum-pool-size: 10
-      minimum-idle: 2
-      idle-timeout: 30000        # 30s
-      connection-timeout: 300000 # 300s
-      max-lifetime: 1800000      # 30m
+  config:
+    import: optional:classpath:application-shared-tenant-qa.yaml
 
-  jpa:
-    open-in-view: false
-    show-sql: true
-    hibernate:
-      ddl-auto: none
-    properties:
-      hibernate:
-        ddl-auto: none
-        default_schema: tenant
-        hbm2ddl.create_namespaces: true
-        format_sql: true
-        jdbc.time_zone: UTC
-  flyway:
-    enabled: true
-    baseline-on-migrate: true
-    locations: classpath:db/migration/common,classpath:db/migration/{vendor}
-    schemas: public,tenant
-    default-schema: tenant
-  kafka:
-    bootstrap-servers: kafka:9092
-    client-id: ejada-tenant-qa
-    consumer:
-      group-id: ejada-backend-qa
-      auto-offset-reset: earliest
-      enable-auto-commit: false
-      properties:
-        max.poll.interval.ms: 300000
-        max.poll.records: 500
-    producer:
-      acks: all
-      retries: 3
-      batch-size: 16384
-      compression-type: gzip
-      properties:
-        linger.ms: 5
-
-
-  data:
-    redis:
-      repositories:
-        enabled: false
+app:
+  service-code: tenant
+  schema: tenant
+  subscription-approval:
+    topic: tenant.subscription-approvals
+    approval-role: ejada-officer
+    consumer-group: tenant-approval-listener
 
 management:
   endpoints:
@@ -79,93 +35,6 @@ springdoc:
     tagsSorter: alpha
 
 shared:
-  core:
-    correlation:
-      header-name: X-Correlation-Id
-      generate-if-missing: true
-      echo-response-header: true
-    tenant:
-      enabled: true
-      header-name: X-Tenant-Id
-      query-param: tenantId
-      default-policy: OPTIONAL
-      echo-response-header: true
-    cors:
-      enabled: true
-      allowed-origins: http://localhost:3000
-      allowed-methods: GET,POST,PUT,DELETE,OPTIONS
-      allowed-headers: "*"
-  headers:
-    enabled: true
-
-    correlation:
-      header: X-Correlation-Id
-      auto-generate: true
-      mandatory: true
-    request:
-      header: X-Request-ID
-      auto-generate: true
-      mandatory: true
-    tenant:
-      header: X-Tenant-Id
-      auto-generate: false
-      mandatory: false
-    user:
-      header: X-User-Id
-      auto-generate: false
-      mandatory: false
-    mdc:
-      enabled: true
-    security:
-      enabled: true
-      hsts:
-        enabled: true
-        max-age: 31536000
-        include-subdomains: true
-      frame-options: SAMEORIGIN
-      content-type-options: nosniff
-      referrer-policy: no-referrer
-      xss-protection: "1; mode=block"
-    propagation:
-      enabled: true
-      include:
-        - X-Correlation-Id
-        - X-Request-ID
-        - X-Tenant-Id
-        - X-User-Id
-  audit:
-    enabled: true
-    web:
-      enabled: true
-      include-headers: true
-      track-bodies: true
-    aop:
-      enabled: true
-    dispatcher:
-      async: true
-    retention:
-      enabled: true
-      days: 90
-    masking:
-      enabled: true
-      fields-by-key:
-        - password
-        - secret
-        - token
-        - ssn
-    sinks:
-      db:
-        enabled: true
-        schema: tenant
-        table: audit_logs
-      kafka:
-        enabled: true
-        topic: audit_logs
-      otlp:
-        enabled: false
-      outbox:
-        enabled: false
-
   openapi:
     enabled: true
     title: "Ejada - tenant Service API"
@@ -177,42 +46,7 @@ shared:
       name: tenant
       packages-to-scan:
         - com.ejada.tenant
-
     jwt-security: true
-  redis:
-    host: redis
-    port: 6379
-    timeout: 2s
-    key-prefix: ejada
-    default-ttl: 600s
-    reactive: false
-  crypto:
-    algorithm: AES_GCM
-    in-memory:
-      activeKid: ${CRYPTO_ACTIVE_KID:local-qa-key}
-      keys:
-        local-qa-key: ${CRYPTO_LOCAL_QA_KEY}
-  security:
-    mode: hs256
-    jwt:
-      secret: ${JWT_SECRET}
-      token-period: 15m
-    hs256:
-      secret: ${JWT_SECRET}
-    resource-server:
-      enabled: true
-      permit-all: ["**"]
-      disable-csrf: false
-    stateless: true
-    roles-claim: roles
-    tenant-claim: tenant
-    enable-role-check: true
-
-app:
-  subscription-approval:
-    topic: tenant.subscription-approvals
-    approval-role: ejada-officer
-    consumer-group: tenant-approval-listener
 
 logging:
   level:
